+++ src/include/fast_alltoall/alltoall_algorithm.h
@@ -21,7 +21,7 @@ struct map_data_t{
 };
 
 struct PermutationSet{
-    uint frequency;
+    uint64_t frequency;
     uint scaling_factor;
     uint dim;
     struct map_data_t mp[MAX_GPU_PER_SERVER]; // mapping from row vertice to col vertice, both indexed from 0-dim-1
@@ -31,9 +31,9 @@ struct PermutationSet{
 void map_insert(struct map_data_t * array, uint * sz, uint key, uint val);
 bool map_lookup(struct map_data_t * array, uint sz, uint key, uint * val);
 
-void init_permutation_set(struct PermutationSet * ps, uint _freq = 1, uint _sf = 1, uint _dim = 0);
-void set_freq_permutation_set(struct PermutationSet * ps, uint freq);
-uint get_freq_permutation_set(struct PermutationSet * ps);
+void init_permutation_set(struct PermutationSet * ps, uint64_t _freq = 1, uint _sf = 1, uint _dim = 0);
+void set_freq_permutation_set(struct PermutationSet * ps, uint64_t freq);
+uint64_t get_freq_permutation_set(struct PermutationSet * ps);
 void to_server_permutation_set(struct PermutationSet * ps, uint server_n, uint * r);
 void from_server_permutation_set(struct PermutationSet * ps, uint server_n, uint * r);
 void print_permutation_set(struct PermutationSet * ps);
